refactor(runbotics-orchestrator-ui): :truck: Changes files structure for ProcessVariablePanel
