+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/ProcessVariablesPanel/VariablesPanel/index.tsx
@@ -2,6 +2,6 @@ import VariablePanel from './VariablesPanel';
 
 export default VariablePanel;
 export { default as VariableRow } from './VariableRow';
-export { default as VariableCopyMenu } from '../VariableCopyMenu';
-export * from './VariableRow.types';
+export { default as VariableCopyMenu } from '../VariableMenu/VariableCopyMenu';
+export * from '../ProcessVariablesPanel.types';
 export * from './VariablesPanel.styles';
